refactor(geo): move detect epsg to projection utils
