+++ packages/geo/src/lib/import-export/shared/drop-geo-file.directive.ts
@@ -17,6 +17,7 @@ import { Feature } from '../../feature/shared/feature.interfaces';
 import { LayerService } from '../../layer/shared/layer.service';
 import { MapBrowserComponent } from '../../map/map-browser/map-browser.component';
 import { IgoMap } from '../../map/shared/map';
+import { detectFileEPSG } from '../../map/shared/projection.utils';
 import { StyleListService } from '../../style/style-list/style-list.service';
 import { StyleService } from '../../style/style-service/style.service';
 import {
@@ -85,7 +86,7 @@ export class DropGeoFileDirective
 
   private onFilesDropped(files: File[]) {
     for (const file of files) {
-      this.detectEPSG(file);
+      detectFileEPSG({ file, epsgCode$: this.epsgCode$ });
       this.epsgCode$$.push(
         this.epsgCode$
           .pipe(
@@ -103,52 +104,6 @@ export class DropGeoFileDirective
           )
       );
     }
-  }
-
-  private detectEPSG(file: File, nbLines: number = 500) {
-    if (
-      !file.name.toLowerCase().endsWith('.geojson') &&
-      !file.name.toLowerCase().endsWith('.gml')
-    ) {
-      this.epsgCode$.next('epsgNotDefined');
-      return;
-    }
-    const reader = new FileReader();
-    reader.onload = (e) => {
-      if (file.name.toLowerCase().endsWith('.geojson')) {
-        const geojson = JSON.parse(reader.result as string);
-        if (geojson.crs?.properties?.name) {
-          const epsg = geojson.crs.properties.name.match(/EPSG:{1,2}\d{0,6}/gm);
-          if (epsg !== null && epsg.length) {
-            this.epsgCode$.next(epsg[0].replace(/::/g, ':'));
-            return;
-          } else {
-            this.epsgCode$.next('epsgNotDefined');
-            return;
-          }
-        } else {
-          this.epsgCode$.next('epsgNotDefined');
-          return;
-        }
-      } else if (file.name.toLowerCase().endsWith('.gml')) {
-        const text = reader.result as string;
-        const lines = (text as string).split('\n');
-        for (let line = 0; line <= nbLines; line++) {
-          const epsg = lines[line].match(/EPSG:\d{0,6}/gm);
-          if (epsg !== null && epsg.length) {
-            this.epsgCode$.next(epsg[0]);
-            break;
-          } else {
-            this.epsgCode$.next(undefined);
-            return;
-          }
-        }
-      } else {
-        this.epsgCode$.next('epsgNotDefined');
-        return;
-      }
-    };
-    reader.readAsText(file, 'UTF-8');
   }
 
   private onFileImportSuccess(file: File, features: Feature[]) {
